refactor(transparent-proxy): get rid of unnecessary struct
There is no V2 or V1 iptables engines. There is only one, so we
don't have to have interface representing engine and can just use
pure functions, so I could get rid of `TransparentProxyV2` struct
and `TransparentProxy` interface.
